First loop closed.

+++ app/src/main/java/org/wikipedia/offline/OfflineObject.kt
@@ -3,5 +3,5 @@ package org.wikipedia.offline
 import java.util.*
 
 data class OfflineObject(val url: String, val lang: String, val path: String, var status: Int) {
-    val usedBy: List<Int> = ArrayList()
+    val usedBy: List<Long> = ArrayList()
 }
